amount, account balance data type changed to float

+++ db/migrate/20200216183503_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[6.0]
 
       t.string :username, null: false
       t.string :name, default: '', null: false
-      t.bigint :account_balance, default: 0, null: false
+      t.float :account_balance, default: 0, null: false
 
       t.timestamps
     end
